apps/streamlit_app/requirements.txt
streamlit
pandas
pyarrow
db-dtypes
google-cloud-bigquery
google-auth
python-dotenv
dbt-core
dbt-duckdb
duckdb
